restructured pet info page and started to add functionality to it. also connected register server and client
